Changed hierarchy to make layout be used as a component
