+++ frontend/sheet-manager/pages/characters/new/index.jsx
@@ -12,9 +12,13 @@ import Overview from './overview'
 import { AnimatePresence } from 'framer-motion'
 import useLocalStorageState from '@/src/utilitaryHooks/use_local_storage_state'
 import useLogin from 'hooks/use_login'
+import MainLayout from '@/src/layouts/main_layout'
+import { useRouter } from 'next/router'
 
 export function Views () {
-  const { parentViewVisibility, subViewVisibility, selectedSubView } = useContext(ScreenSlideContext)
+  const router = useRouter()
+
+  const { parentViewVisibility, subViewVisibility, selectedSubView, setParentViewVisibility, setSubViewVisibility } = useContext(ScreenSlideContext)
 
   const [character, setCharacter] = useLocalStorageState('new_character_v1', {
     level: 1,
@@ -57,11 +61,23 @@ export function Views () {
   }
 
   return (
-    <AnimatePresence initial={false}>
-      {parentViewVisibility && <Overview character={character} setCharacter={setCharacter} />}
-      {subViewVisibility && renderSubView()}
+    <MainLayout onLayoutBack={e => {
+      if (parentViewVisibility) {
+        router.back()
+      } else {
+        setCharacter({
+          ...character,
+        })
+        setParentViewVisibility(true)
+        setSubViewVisibility(false)
+      }
+    }}>
+      <AnimatePresence initial={false}>
+        {parentViewVisibility && <Overview character={character} setCharacter={setCharacter} />}
+        {subViewVisibility && renderSubView()}
 
-    </AnimatePresence>
+      </AnimatePresence>
+    </MainLayout>
   )
 }
