+++ frontend/sheet-manager/src/components/container.jsx
@@ -1,7 +1,7 @@
 export default function Container({ children }) {
-    return (
-        <div className="px-7 sm:px-10 md:px-20 lg:px-32">
-            {children}
-        </div>
-    )
+  return (
+    <div className="px-7 sm:px-10 md:px-20 lg:px-32">
+      {children}
+    </div>
+  )
 }
